zmiana tempa z shiftem * 0.1

+++ interface/songEditor/songEditor.h
@@ -91,7 +91,7 @@ public:
 	void markCurrentPattern(uint8_t forceRefresh);
 	void readSong();
 
-	void changeGlobalTempo(int16_t value);
+	void changeGlobalTempo(float value);
 	void setDefaultGlobalTempo();
